Refactored the variable noise generator to align better with existing pycbc functionality, and moved the example to a seperate script in the example folder.
